Moved join and split to the utility namespace
