+++ src/elementary/ParticlePairID/src/validate.hpp
@@ -5,13 +5,13 @@
  */
 static bool validate( const std::string& string ) {
 
-  auto strings = split( string, "," );
+  auto strings = utility::split( string, "," );
   if ( strings.size() != 2 ) {
 
     if ( ReactionType::isRegistered( string ) ) {
 
       ReactionType type( string );
-      if ( ( type.isSpecial() or ( type.name() == "capture" ) ) and 
+      if ( ( type.isSpecial() or ( type.name() == "capture" ) ) and
            ( type.name() != "elastic" ) ) {
 
         return true;
