+++ src/elementary/test/elementary.test.cpp
@@ -4,8 +4,6 @@
 #include "elementary/src/absorb.hpp"
 #include "elementary/src/emit.hpp"
 #include "elementary/src/resolve.hpp"
-#include "elementary/src/join.hpp"
-#include "elementary/src/split.hpp"
 #include "elementary/src/fromEndfReactionNumber.hpp"
 #include "elementary/src/toEndfReactionNumber.hpp"
 
@@ -23,7 +21,5 @@ using namespace njoy::elementary;
 #include "elementary/test/absorb.test.hpp"
 #include "elementary/test/emit.test.hpp"
 #include "elementary/test/resolve.test.hpp"
-#include "elementary/test/join.test.hpp"
-#include "elementary/test/split.test.hpp"
 #include "elementary/test/fromEndfReactionNumber.test.hpp"
 #include "elementary/test/toEndfReactionNumber.test.hpp"
