+++ src/utility/test/utility.test.cpp
@@ -1,0 +1,13 @@
+#define CATCH_CONFIG_MAIN
+
+#include "catch.hpp"
+#include "utility/join.hpp"
+#include "utility/split.hpp"
+
+// other includes
+
+// convenience typedefs
+using namespace njoy::utility;
+
+#include "utility/test/join.test.hpp"
+#include "utility/test/split.test.hpp"
